# Conflicts:
#	.idea/gradle.xml
#	.idea/libraries/appcompat_v7_23_1_1.xml
#	.idea/libraries/support_annotations_23_1_1.xml
#	.idea/libraries/support_v4_23_1_1.xml
#	app/app.iml

+++ app/src/main/java/com/asimkhanal/alarmclock/AlarmActivity.java
@@ -19,14 +19,15 @@ import java.util.List;
 import java.util.Random;
 
 public class AlarmActivity extends Activity implements SensorListener {
-    AlarmManager alarmManager;
-    AudioManager audioManager;
-    private PendingIntent pendingIntent;
+    private static final int SHAKE_DURATION =100000 ;
+    private AlarmManager alarmManager;
+    private AudioManager audioManager;
     private static final String LOG_TAG = "callIntent";
-    SensorManager sensorMgr;
-    private static final int SHAKE_THRESHOLD = 1600;
+    private SensorManager sensorManager;
+    private static final int SHAKE_THRESHOLD = 3200;
     private long lastUpdate;
-    private float x,y,z,last_x,last_y,last_z;
+    private float last_x, last_y, last_z;
+    private long mLastShake;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -35,8 +36,8 @@ public class AlarmActivity extends Activity implements SensorListener {
         setTitle("Alarm Actions");
         alarmManager = (AlarmManager) getSystemService(ALARM_SERVICE);
         audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        sensorMgr = (SensorManager)getSystemService(SENSOR_SERVICE);
-        sensorMgr.registerListener(this,
+        sensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
+        sensorManager.registerListener(this,
                 SensorManager.SENSOR_ACCELEROMETER,
                 SensorManager.SENSOR_DELAY_GAME);
         lastUpdate = System.currentTimeMillis();
@@ -45,9 +46,11 @@ public class AlarmActivity extends Activity implements SensorListener {
         //audioManager.setMicrophoneMute(true);
     }
 
-    public void stopButtonClicked(View V){
-        android.os.Process.killProcess(android.os.Process.myPid());
-    }
+// --Commented out by Inspection START (4/2/16, 12:05 AM):
+//    public void stopButtonClicked(View V){
+//        android.os.Process.killProcess(android.os.Process.myPid());
+//    }
+// --Commented out by Inspection STOP (4/2/16, 12:05 AM)
 
 //    private class EndCallListener extends PhoneStateListener {
 //        @Override
@@ -73,23 +76,23 @@ public class AlarmActivity extends Activity implements SensorListener {
 
         DatabaseHelper db = DatabaseHelper.getInstance(this);
         List<Contact> contacts = db.getContactsByTier(tier);
-        Contact contactToCall = contacts.get(new Random().nextInt(contacts.size()));
+        Contact contactToCall = contacts.get((new Random()).nextInt(contacts.size()));
 
 
         // May need to run following block in service
 //        EndCallListener callListener = new EndCallListener();
 //        TelephonyManager mTM = (TelephonyManager) getSystemService(TELEPHONY_SERVICE);
 //        mTM.listen(callListener, PhoneStateListener.LISTEN_CALL_STATE);
-
+        Log.d("atag",contactToCall.phone_number);
         String contactNumber = "tel:" + contactToCall.phone_number;
 
         Log.d("MyActivity", "Alarm On");
         Calendar calendar = Calendar.getInstance();
         calendar.add(Calendar.MINUTE, 1);
         Intent myIntent = new Intent(AlarmActivity.this, AlarmReceiver.class);
-        pendingIntent = PendingIntent.getBroadcast(AlarmActivity.this, 0, myIntent, 0);
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(AlarmActivity.this, 0, myIntent, 0);
         alarmManager.set(AlarmManager.RTC, calendar.getTimeInMillis(), pendingIntent);
-        Log.d("AddActivity", "alarm");
+        Log.d("AddAlarmActivity", "alarm");
 
         Intent callIntent = new Intent(Intent.ACTION_CALL);
         callIntent.setData(Uri.parse(contactNumber));
@@ -118,16 +121,19 @@ public class AlarmActivity extends Activity implements SensorListener {
                 long diffTime = (curTime - lastUpdate);
                 lastUpdate = curTime;
 
-                x = values[SensorManager.DATA_X];
-                y = values[SensorManager.DATA_Y];
-                z = values[SensorManager.DATA_Z];
+                float x = values[SensorManager.DATA_X];
+                float y = values[SensorManager.DATA_Y];
+                float z = values[SensorManager.DATA_Z];
 
-                float speed = Math.abs(x+y+z - last_x - last_y - last_z) / diffTime * 10000;
+                float speed = Math.abs(x + y + z - last_x - last_y - last_z) / diffTime * 10000;
 
-                if (speed > SHAKE_THRESHOLD) {
-                    Log.d("sensor", "shake detected w/ speed: " + speed);
-                    Toast.makeText(this, "shake detected w/ speed: " + speed, Toast.LENGTH_SHORT).show();
-                    android.os.Process.killProcess(android.os.Process.myPid());
+                if (speed > SHAKE_THRESHOLD ) {
+                    if(curTime-mLastShake>SHAKE_DURATION) {
+                        mLastShake=curTime;
+                        Log.d("sensor", "shake detected w/ speed: " + speed);
+                        Toast.makeText(this, "shake detected w/ speed: " + speed, Toast.LENGTH_SHORT).show();
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                    }
                 }
                 last_x = x;
                 last_y = y;
